fix: LazyInitialization 오류 해결
