+++ src/main/resources/application-local.yml
@@ -1,0 +1,40 @@
+spring:
+  h2:
+    console:
+      enabled: false
+      path: /h2-console
+
+  application:
+    name: course-registration
+
+  datasource:
+    url: jdbc:mysql://${DB_HOST}:${DB_PORT}/${DB_NAME}
+    username: ${DB_USERNAME}
+    password: ${DB_PASSWORD}
+    driver-class-name: com.mysql.cj.jdbc.Driver
+
+  jpa:
+    hibernate:
+      ddl-auto: update   # 토이면 update 또는 create-drop
+    database-platform: org.hibernate.dialect.MySQL8Dialect
+    properties:
+      hibernate:
+        format_sql: true
+
+  data:
+    redis:
+      host: localhost
+      port: 6379
+      connect-timeout: 2000
+      database: 0
+
+  #  mvc:
+  #    throw-exception-if-no-handler-found: true
+  web:
+    resources:
+      add-mappings: false
+
+server:
+  servlet:
+    session:
+      timeout: 30m # 마지막 활동으로부터 30분 지나면 로그아웃
